Refactored expected / received actions mismatch message
This refactors the display we use to show the mismatch between the
expected and the received actions in `listenForActions`. Now instead of
showing the full list of the expected actions and the full list of
received actions we display the symmetrical difference (A - B and B - A)
to let the user know what's up more quickly.